Shell escape the value of Elastic APM custom properties
The current behavior is for no escaping to occur. This if you use a value like `"object_name[java.lang:type=Memory] attribute[HeapMemoryUsage:metric_name=test_heap_metric]"` (as we see in #786), the literal value is inserted into a system property and the space (among other characters) breaks the start command.

This change will use `Shellwords.escape(..)` the value for every custom property unless that property value contains what looks like a subshell `$(..)` or environment variable `${..}` reference. If it looks like a subshell or env variable is being referenced, we will not escape just wrap the value in quotes. We wrap it in quotes in case the shell variable or subshell returns something which includes spaces. This is not perfect though, and you need to be careful if using subshell/env variables.

For example:

- `object_name[java.lang:type=Memory] attribute[HeapMemoryUsage:metric_name=test_heap_metric]` becomes `object_name\[java.lang:type\=Memory\]\ attribute\[HeapMemoryUsage:metric_name\=test_heap_metric\]`
- `$(echo "Hello world!") and stuff` becomes `"$(echo "Hello world!") and stuff"
- `--> ${SOME_VAR} <--` becomes `"--> ${SOME_VAR} <--"`

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>